Tool call wip

+++ orchestration/src/main/java/com/sap/ai/sdk/orchestration/spring/OrchestrationSpringChatResponse.java
@@ -87,8 +87,6 @@ public class OrchestrationSpringChatResponse extends ChatResponse {
   @Nonnull
   private static DefaultUsage toDefaultUsage(@Nonnull final TokenUsage usage) {
     return new DefaultUsage(
-        usage.getPromptTokens().longValue(),
-        usage.getCompletionTokens().longValue(),
-        usage.getTotalTokens().longValue());
+        usage.getPromptTokens(), usage.getCompletionTokens(), usage.getTotalTokens());
   }
 }
